requirements.txt
Django==2.1.2
django-debug-toolbar==2.0
django-filter==2.2.0
djangorestframework==3.10.3
pytz==2019.3
sqlparse==0.3.0
whitenoise==4.1.4
